fix colorbar

+++ app/javascript/controllers/suppression_controller.js
@@ -1,7 +1,7 @@
 import { Controller } from "@hotwired/stimulus";
 
 export default class extends Controller {
-  static targets = ["card", "checkbox"];
+  static targets = ["card", "checkbox", "deleteButton", "cancelButton", "listeElements"];
 
   connect() {
     console.log('Connecté à suppression');
@@ -31,34 +31,37 @@ export default class extends Controller {
 
     const token = document.querySelector('meta[name="csrf-token"]').content;
 
-    fetch(url, {
-      method: 'DELETE',
-      headers: {
-        'Content-Type': 'application/json',
-        'X-CSRF-Token': token
-      },
-      body: JSON.stringify({ pokemon_ids: selectedPokemonIds })
-    }).then(response => {
-      if (response.ok) {
-        console.log('Cartes supprimées avec succès');
-        // Recharger la partie de la page contenant la liste des éléments
-        fetch(window.location.href)
-          .then(response => response.text())
-          .then(data => {
-            const parser = new DOMParser();
-            const htmlDocument = parser.parseFromString(data, 'text/html');
-            const updatedListeElements = htmlDocument.getElementById('liste-elements');
-            const listeElementsDiv = document.getElementById('liste-elements');
-            listeElementsDiv.innerHTML = updatedListeElements.innerHTML;
-          })
-          .catch(error => console.error('Erreur lors du chargement de la partie mise à jour de la page :', error));
-      } else {
-        console.error('Erreur lors de la suppression des cartes');
-      }
-    }).catch(error => {
-      console.error('Erreur lors de la suppression des cartes:', error);
-    });
-  }
+    // Confirmation avant la suppression
+    const confirmation = confirm("Êtes-vous sûr de vouloir supprimer les cartes sélectionnées ?");
+    if (confirmation) {
+      fetch(url, {
+        method: 'DELETE',
+        headers: {
+          'Content-Type': 'application/json',
+          'X-CSRF-Token': token
+        },
+        body: JSON.stringify({ pokemon_ids: selectedPokemonIds })
+      }).then(response => {
+        if (response.ok) {
+          console.log('Cartes supprimées avec succès');
+          // Recharger uniquement la partie de la page contenant la liste des éléments
+          fetch(window.location.href)
+            .then(response => response.text())
+            .then(data => {
+              const parser = new DOMParser();
+              const htmlDocument = parser.parseFromString(data, 'text/html');
+              const updatedListeElements = htmlDocument.getElementById('liste-elements').innerHTML;
+              this.listeElementsTarget.innerHTML = updatedListeElements;
+            })
+            .catch(error => console.error('Erreur lors du chargement de la partie mise à jour de la page :', error));
+        } else {
+          console.error('Erreur lors de la suppression des cartes');
+        }
+      }).catch(error => {
+        console.error('Erreur lors de la suppression des cartes:', error);
+      });
+    }
+}
 
   toggleSelection(event) {
     const checkbox = event.currentTarget;
